+++ RapidApi/RapidApi.Cli.csproj
@@ -10,14 +10,6 @@
   </PropertyGroup>
 
   <ItemGroup>
-    <None Remove="Connected Services\RapidApi Service\RapidApi ServiceCsdl.xml" />
-  </ItemGroup>
-
-  <ItemGroup>
-    <EmbeddedResource Include="Connected Services\RapidApi Service\RapidApi ServiceCsdl.xml" />
-  </ItemGroup>
-
-  <ItemGroup>
     <PackageReference Include="Azure.Identity" Version="1.3.0" />
     <PackageReference Include="Azure.Storage.Files.Shares" Version="12.5.0" />
     <PackageReference Include="Ductus.FluentDocker" Version="2.7.3" />
